% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AIC.yppe}
\alias{AIC.yppe}
\title{Akaike information criterion}
\usage{
\method{AIC}{yppe}(object, ..., k = 2)
}
\arguments{
\item{object}{an object of the class yppe.}

\item{...}{further arguments passed to or from other methods.}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC.}
}
\value{
the Akaike information criterion
}
\description{
Akaike information criterion
}
