% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.yppe}
\alias{anova.yppe}
\title{anova method for yppe models}
\usage{
\method{anova}{yppe}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}
}
\value{
a list containing the estimated survival probabilities.
}
\description{
Compute analysis of variance (or deviance) tables for one or more fitted model objects.
}
