% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.yppe}
\alias{coef.yppe}
\title{Estimated regression coefficients}
\usage{
\method{coef}{yppe}(object, ...)
}
\arguments{
\item{object}{an object of the class yppe}

\item{...}{further arguments passed to or from other methods}
}
\value{
the estimated regression coefficients
}
\description{
Estimated regression coefficients
}
