% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexp.R
\name{findInt}
\alias{findInt}
\title{Function to identify the times' intervals}
\usage{
findInt(time, tgrid)
}
\arguments{
\item{time}{vector of times}

\item{tgrid}{time grid}
}
\value{
indicator function of times's intervals
}
\description{
Function to identify the times' intervals
}
