% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{gastric}
\alias{gastric}
\title{Gastric data set}
\description{
Gastric data set
}
\references{
\url{data_blah.com}
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{data}
