% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLik.yppe}
\alias{logLik.yppe}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{yppe}(object, ...)
}
\arguments{
\item{object}{an object of the class yppe.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the log-likelihood associated with the fitted model.
}
\description{
Extract Log-Likelihood
}
