% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model.matrix.yppe}
\alias{model.matrix.yppe}
\title{Model.matrix method for yppe models}
\usage{
\method{model.matrix}{yppe}(object, ...)
}
\arguments{
\item{object}{an object of the class yppe.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The model matrix (or matrices) for the fit.
}
\description{
Reconstruct the model matrix (or matrices if the alternative formulation of the YP model is used) for a yppe model.
}
\examples{
\dontrun{
fit <- yppe(Surv(time, status)~arm, data=ipass)
model.matrix(fit)
}

}
