% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ovarian}
\alias{ovarian}
\title{Ovarian data set}
\description{
Ovarian data set
}
\references{
\url{data_blah.com}
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{data}
