% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexp.R
\name{pehaz}
\alias{pehaz}
\alias{hpexp}
\alias{Hpexp}
\title{Hazard and cumulative hazard functions of the PE distribution}
\usage{
hpexp(x, rho, rates)

Hpexp(x, rho, rates)
}
\arguments{
\item{x}{vector of time points.}

\item{rho}{vector of time grid knots.}

\item{rates}{vector of failure rates.}
}
\value{
hpexp gives the hazard function and Hpexp gives the cumulative hazard function of the PE distribution.
}
\description{
Hazard and cumulative hazard functions of the PE distribution
}
