% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_phpe.R
\name{print.summary.phpe}
\alias{print.summary.phpe}
\title{Print the summary.phpe output}
\usage{
\method{print}{summary.phpe}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.phpe.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.phpe output
}
