% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_pope.R
\name{print.summary.pope}
\alias{print.summary.pope}
\title{Print the summary.pope output}
\usage{
\method{print}{summary.pope}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.pope.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.pope output
}
