% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_yppe.R
\name{print.summary.yppe}
\alias{print.summary.yppe}
\title{Print the summary.yppe output}
\usage{
\method{print}{summary.yppe}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.yppe.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.yppe output
}
