% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rates}
\alias{rates}
\title{Generic S3 method rates}
\usage{
rates(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the estimated failure rates for the PE distribution.
}
\description{
Generic S3 method rates
}
\details{
Method only available for ML approach.
}
