% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rates.yppe}
\alias{rates.yppe}
\title{Estimated failure rates for the PE distribution}
\usage{
\method{rates}{yppe}(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the estimated failure rates for the PE distribution.
}
\description{
Estimated failure rates for the PE distribution
}
\details{
Method only available for ML approach.
}
