% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{rsurv}
\alias{rsurv}
\title{Random generation of survival data}
\usage{
rsurv(
  formula,
  covariates,
  baseline = "weibull",
  gamma,
  psi = NULL,
  phi = NULL,
  nu = NULL,
  max_fu
)
}
\arguments{
\item{formula}{formula specifying the linear predictors}

\item{covariates}{data frame containing the covariates used to generate the survival times}

\item{baseline}{baseline model (currently only the Weibull distribution is available)}

\item{gamma}{baseline parameters}

\item{psi}{short-term regression coefficients}

\item{phi}{long-term regression coefficients}

\item{max_fu}{maximum follow-up time}
}
\description{
Function to generate a random sample of survival data.
}
