% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexp.R
\name{sturges}
\alias{sturges}
\title{Sturges rule to computing the number of intervals}
\usage{
sturges(n)
}
\arguments{
\item{n}{Sample size.}
}
\value{
The number of intervals \code{n}.
}
\description{
Sturges rule to computing the number of intervals
}
\examples{
sturges(100)
}
