% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_phpe.R
\name{summary.phpe}
\alias{summary.phpe}
\title{Summary for the yppe model}
\usage{
\method{summary}{phpe}(object, ...)
}
\arguments{
\item{object}{an objecto of the class 'yppe'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary for the yppe model
}
