% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_phpe.R
\name{survfit.phpe}
\alias{survfit.phpe}
\title{survfit method for phpe models}
\usage{
\method{survfit}{phpe}(formula, newdata, ...)
}
\arguments{
\item{formula}{an object of the class phpe}

\item{newdata}{a data frame containing the set of explanatory variables.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list containing the estimated survival probabilities.
}
\description{
Computes the predicted survivor function for a phpe model.
}
\examples{
\donttest{
# ML approach:
library(YPPE)
mle <- phpe(Surv(time, status)~arm, data=ipass, approach="mle", init = 0)
summary(mle)
ekm <- survival::survfit(Surv(time, status)~arm, data=ipass)
newdata <- data.frame(arm=0:1)
St <- survfit(mle, newdata)
plot(ekm, col=1:2)
with(St, lines(time, surv[[1]]))
with(St, lines(time, surv[[2]], col=2))

# Bayesian approach:
bayes <- phpe(Surv(time, status)~arm, data=ipass, approach="bayes")
summary(bayes)
ekm <- survival::survfit(Surv(time, status)~arm, data=ipass)
newdata <- data.frame(arm=0:1)
St <- survfit(bayes, newdata)
plot(ekm, col=1:2)
with(St, lines(time, surv[[1]]))
with(St, lines(time, surv[[2]], col=2))
}

}
