% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexp.R
\name{timeGrid}
\alias{timeGrid}
\title{Time grid}
\usage{
timeGrid(time, status, n.int = NULL)
}
\arguments{
\item{time}{Vector of failure times}

\item{status}{Vector of failure indicators}

\item{n.int}{Optional. Number of intervals. If \code{NULL}, the number of intervals is
set to be equal to the number of distinct observed failure times.}
}
\value{
Time grid.
}
\description{
Time grid
}
