% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcov}
\alias{vcov}
\title{Generic S3 method vcov}
\usage{
vcov(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the variance-covariance matrix associated the regression coefficients.
}
\description{
Generic S3 method vcov
}
