% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcov.yppe}
\alias{vcov.yppe}
\alias{vcov.ypppe}
\title{Variance-covariance matrix for a yppe model}
\usage{
\method{vcov}{yppe}(object, ...)
}
\arguments{
\item{object}{an object of the class yppe.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the variance-covariance matrix associated with the regression coefficients.
}
\description{
This function extracts and returns the variance-covariance matrix associated with the regression coefficients when the maximum likelihood estimation approach is used in the model fitting.
}
