% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_yppe.R
\name{crossTime}
\alias{crossTime}
\alias{crossTime.yppe}
\title{Function to compute the crossing survival times}
\usage{
crossTime(object, ...)

\method{crossTime}{yppe}(object, newdata1, newdata2, prob = c(0.025,
  0.975), nboot = 4000)
}
\arguments{
\item{object}{an object of the class yppe.}

\item{...}{further arguments passed to or from other methods.}

\item{newdata1}{a data frame containing the first set of explanatory variables}

\item{newdata2}{a data frame containing the second set of explanatory variables}

\item{prob}{quantiles used to compute the confidence/credible intervals}

\item{nboot}{number of bootstrap samples (default nboot=4000); ignored if approach="bayes".}
}
\value{
the crossing survival time
}
\description{
Function to compute the crossing survival times

Function to compute the crossing survival times
}
