% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_yppe.R
\name{survfit}
\alias{survfit}
\alias{survfit.yppe}
\title{Survival function of the yppe model (original formulation)}
\usage{
survfit(object, ...)

\method{survfit}{yppe}(object, newdata)
}
\arguments{
\item{object}{an object of the class yppe.}

\item{...}{further arguments passed to or from other methods.}

\item{newdata}{a data frame containing the set of explanatory variables.}
}
\value{
a list containing the estimated survival probabilities.
}
\description{
Survival function of the yppe model (original formulation)

Survival function of the yppe model (original formulation)
}
