% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_yppe.R
\name{survfit}
\alias{survfit}
\title{Generic S3 method}
\usage{
survfit(object, ...)
}
\arguments{
\item{object}{an object of class yppe}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic S3 method
}
