% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YPPE-package.R
\docType{package}
\name{YPPE-package}
\alias{YPPE-package}
\alias{YPPE}
\title{The 'YPPE' package.}
\description{
Semiparametric modeling of lifetime data with crossing survival curves via Yang and Prentice model with piecewise exponential baseline distribution curves. Model fitting can be carried out by either likelihood-based or Bayesian approaches. The package interfaces with Stan to fit the YPPE model. It also provides a function to obtain point and interval estimates for the crossing survival times.
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org

Yang, S. and Prentice, R. L. (2005). Semiparametric analysis of short-term and long-term hazard ratios with two-sample survival data. Biometrika 92, 1-17.
}
